[Pal] Introduce {read,write}_all wrappers
These are partial-read/write-safe wrappers for read/write syscalls.

This commit also fixes a rare bug on process creation, where data didn't
fit into a single write.

+++ Pal/src/host/Skeleton/db_streams.c
@@ -38,6 +38,6 @@ int _DkInitDebugStream(const char* path) {
     return -PAL_ERROR_NOTIMPLEMENTED;
 }
 
-ssize_t _DkDebugLog(const void* buf, size_t size) {
+int _DkDebugLog(const void* buf, size_t size) {
     return -PAL_ERROR_NOTIMPLEMENTED;
 }
