+++ Pal/src/pal_internal.h
@@ -298,7 +298,7 @@ void free(void* mem);
 #endif
 
 int _DkInitDebugStream(const char* path);
-ssize_t _DkDebugLog(const void* buf, size_t size);
+int _DkDebugLog(const void* buf, size_t size);
 void _DkPrintConsole(const void* buf, size_t size);
 int printf(const char* fmt, ...) __attribute__((format(printf, 1, 2)));
 int vprintf(const char* fmt, va_list ap) __attribute__((format(printf, 1, 0)));
